test(bdd): align metadata/raw capture assertions; clippy tidy
