+++ tests/bdd/Cargo.toml
@@ -28,6 +28,7 @@ serde_json.workspace = true
 tempfile.workspace = true
 anyhow.workspace = true
 tokio.workspace = true
+base64.workspace = true
 
 [features]
 default = []
