+++ tests/bdd/steps/golden_fixtures.rs
@@ -1,0 +1,2 @@
+// Golden fixtures step definitions are handled by existing steps.
+// This module exists for organizational purposes and future golden-fixture-specific steps.
